requirements.txt
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==80.10.2
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
tzdata==2025.3
